Moved Flusher into Cache package
